+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 * [HexaPDF::Layout::ColumnBox], [HexaPDF::Layout::ListBox] and
   [HexaPDF::Layout::ImageBox] to correctly respond to `#empty?`
+* [HexaPDF::Layout::ListBox] to take different final box position into account
 * [HexaPDF::Content::Canvas#text] to set the leading only when multiple lines
   are drawn